move/rename Grub tool to GrubConfig in its own file.
